feat: add variant metrics (#253)
## Description

This change adds most variant metrics functionality to the Python SDK.
Variant metrics should be counted whenever get_variant is called:

- if a feature doesn't have variants, register the disabled variant
- if a feature is disabled, register the disabled variant
- otherwise, register the variant that was returned

There is one piece of functionality that isn't included in this PR:

- if a feature doesn't exist, it should still register the disabled variant.

As far as I can tell, this SDK doesn't count metrics for toggles that don't exist, so there was no obvious way to add variant metrics for these features. Both of those cases will be implemented in a follow-up PR.

## Discussion points

### Serialization of emtpy dicts

When serializing metrics, I have done the simple thing and added the variant counts dict, regardless of whether it is empty or not. It might be preferable to not include that key at all if the feature doesn't have any variant metrics. I'd be happy to implement that if we think that's better.

### Tests

I extended the `test_aggregate_and_send_metrics` test with variant data instead of adding a new test. Because variant metrics now are a part of metrics, that felt reasonable. However, we can create a separate test if that's preferable.

### Counting variants on the feature level

The variant counting is currently handled on the feature (`Feature` class) level instead of on the variant (`Variant` class) level.
This was done for a few reasons:
- most importantly: we want to count stats for the disabled variant. Because the disabled variant isn't added as part of the `Feature`'s variants, we'd have to count this separately if we placed counts on the `Variant` level.
- `yes_count` and `no_count` are already tracked on the feature, so it felt natural to count variants there too
- it simplifies serialization and the `reset_stats` method

### Casting `DISABLED_VARIANT["name"]`

Described in the [inline comment](https://github.com/Unleash/unleash-client-python/pull/253#discussion_r1205484173):



This cast appears to be the only way to get the linter to accept that we get a string here. As far as I can tell, "name" should always exist and will always be a string. However, the disabled variant isn't typed in any particular way, so it just shows up as a Dict[str, object].

I'd rather avoid this cast if I can, but I don't know if it's possible without re-architecting significant portions of the code and introducing breaking changes. Input would be much appreciated.



## Type of change

Please delete options that are not relevant.

- [X] New feature (non-breaking change which adds functionality)

# How Has This Been Tested?

Please describe the tests that you ran to verify your changes.

- [X] Unit tests
- [ ] Spec Tests
- [ ] Integration tests / Manual Tests

# Checklist:

- [X] My code follows the style guidelines of this project
- [X] I have performed a self-review of my own code
- [X] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [X] My changes generate no new warnings
- [X] I have added tests that prove my fix is effective or that my feature works
- [X] New and existing unit tests pass locally with my changes (the ETAG test is failing, but it fails on main too)
- [ ] Any dependent changes have been merged and published in downstream modules (I don't think there are any downstream modules?)


## Commits

* Tests: add tests for some variant metrics scenarios

* Chore: make black happy

* tests: fix test logic

* feat: count variants

* feat: reset variant count stats

* Tests: fix

* feat: count variants

* Test: ensure variants are reset on reset_stats

* Tests: make tests test the correct thing

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Test: ensure that variant data is sent correctly

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* feat: add type annotation to variant counts dict

* feat: import Dict typing

* feat: describe param

* Chore: cast str

* Chore: remove comment

* Update tests/unit_tests/test_features.py

Co-authored-by: Simon Hornby <liquidwicked64@gmail.com>

* fix: prefix `_count_variant` with an underscore to "hide" the method

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Simon Hornby <liquidwicked64@gmail.com>