chore: remove duplicated test
This test appears to be the **exact same** as the one right above
it (`test_uc_server_error`): copying one into the other triggers no
changes from git.

Unless I'm missing something, I suggest we remove it because it's both
confusing to have two identical tests, and because it takes 20 seconds
to sleep.