test: skip failing test (#258)
# Description

This PR skips the test that's currently failing until July 1st. We suspect the reason it's failing is because of a dependency and/or something in the scaffolding. It suddenly started failing a few weeks ago. We have spent some hours looking into what's wrong, but haven't been able to fix it just yet.

Please delete options that are not relevant.

## Commits 

* tests: skip failing tests

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* tests: skip test until July

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>