fix: don't count stats for parent flags, even in get_variant (#300)
This PR fixes an issue where parent flags would have metrics counted when you call get_variant on a child flag. This was due to a missing skip_stats argument in the get_variant method of a feature.

I have added a test and the specific data setup that was reported to cause the issue.